==> two_stages_classifier/testing/07_accuracy_consistency_across_genome_stage2.ipynb <==
### Evaluate the GutEuk's performance across different genomes (See EukRep paper, Fig 2C)

import pandas as pd

stage1_res = pd.read_csv("/fs/ess/PAS0439/MING/cilates_fungi_classifier/test_res_stage1.csv")[["seq", "ensembled_out"]]
stage2_res = pd.read_csv("/fs/ess/PAS0439/MING/cilates_fungi_classifier/test_res_stage2.csv")[["seq", "ensembled_out"]]
stage1_res.rename(columns = {"ensembled_out":"stage1_pred"}, inplace = True)
stage2_res.rename(columns = {"ensembled_out":"stage2_pred"}, inplace = True)

seq_origin = pd.read_csv("/fs/ess/PAS0439/MING/cilates_fungi_classifier/testset_seq_origin.csv")

testing_merged_tmp = pd.merge(stage1_res, seq_origin, on = "seq")
testing_merged = pd.merge(testing_merged_tmp, stage2_res, on = "seq", how = "left")

for index, row in testing_merged.iterrows():
    stage1 = row["stage1_pred"]
    stage2 = row["stage2_pred"]
    if stage1 == 0:
        stage1 = "prokaryotes"
        stage2 = "prokaryotes"
    else:
        stage1 = "eukaryotes"
        if stage2 == 0:
            stage2 = "fungi"
        elif stage2 == 1:
            stage2 = "protozoa"
    testing_merged.loc[index, "stage1_pred"] = stage1
    testing_merged.loc[index, "stage2_pred"] = stage2

testing_merged.query("category == 'SAGs'")

fungi_genomes = set(testing_merged.query('category == "fungi"').genome)
protozoa_genomes = set(testing_merged.query('category == "protozoa"').genome) 
sags_genomes =  set(testing_merged.query('category == "SAGs"').genome)
prokaryotes_genomes = set(testing_merged.query('category == "prokaryotes"').genome)

genome_origin = []
accuracy_list = []
genome_list = []

for fungi in fungi_genomes:
    df = testing_merged.query('genome == @fungi')
    correct = 0
    incorrect = 0
    for index, row in df.iterrows():
        pred = row['stage2_pred']
        Y = row['category']
        if pred == Y:
            correct += 1
        else:
            incorrect += 1
    
    accuracy = correct / (correct + incorrect)
    genome_origin.append("fungi")
    accuracy_list.append(accuracy)
    genome_list.append(fungi)
        


for protozoa in protozoa_genomes:
    df = testing_merged.query('genome == @protozoa')
    correct = 0
    incorrect = 0
    for index, row in df.iterrows():
        pred = row['stage2_pred']
        Y = row['category']
        if pred == Y:
            correct += 1
        else:
            incorrect += 1
    
    accuracy = correct / (correct + incorrect)
    genome_origin.append("protozoa")
    accuracy_list.append(accuracy)
    genome_list.append(protozoa)


for sag in sags_genomes:
    df = testing_merged.query('genome == @sag')
    correct = 0
    incorrect = 0
    for index, row in df.iterrows():
        pred = row['stage2_pred']
        Y = row['category']
        if pred == "protozoa":
            correct += 1
        else:
            incorrect += 1
    
    accuracy = correct / (correct + incorrect)
    genome_origin.append("SAG")
    accuracy_list.append(accuracy)
    genome_list.append(sag)


for proka in prokaryotes_genomes:
    df = testing_merged.query('genome == @proka')
    correct = 0
    incorrect = 0
    for index, row in df.iterrows():
        pred = row['stage1_pred']
        Y = row['category']           
        if pred == Y:
            correct += 1
        else:
            incorrect += 1
    
    accuracy = correct / (correct + incorrect)
    genome_origin.append("prokaryotes")
    accuracy_list.append(accuracy)
    genome_list.append(proka)


accuracy_per_genome = pd.DataFrame({"genome_origin": genome_origin, "accuracy": accuracy_list, "genome": genome_list})
accuracy_per_genome.to_csv("/fs/ess/PAS0439/MING/cilates_fungi_classifier/stats_visualization/accuracy_per_genome_stage2.csv", index = None)
